refactor: move networkConfigUrl, rpcUrl
